backend/requirements.txt
# Azure Functions Python Dependencies
# Controle PGM - Backend

# Azure Functions runtime
azure-functions>=1.17.0

# Azure Tables SDK
azure-data-tables>=12.5.0

# Authentication
PyJWT>=2.8.0
bcrypt>=4.1.0

# Validation & Serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Environment
python-dotenv>=1.0.0

# Timezone handling
pytz>=2024.1

# Development dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0
ruff>=0.1.9
